+++ NFTDetection/perceptual_hash_c/ImageHash/DCTII.h
@@ -1,4 +1,4 @@
 #pragma once
 
 // storing DCT in a seprate file, in case speed improved version is desired
-void dctII(double* out_Matrix, double* in_Matrix, size_t columns, size_t rows);
+void dctII(double* out_Matrix, size_t* in_Matrix, size_t columns, size_t rows);
